TI_System/requirements.txt
# File: requirements.txt
# ===========================
#Requirements file for Threat Intelligence System

# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-dateutil>=2.8.2

# Country detection
pycountry>=23.12.0
tldextract>=5.0.0
geoip2>=4.7.0

# NLP (optional)
spacy>=3.7.0

# Database
sqlite3  # Built-in with Python

# OTX Integration
OTXv2>=1.5.12

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
unicodedata2>=15.1.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development (optional)
black>=23.0.0
flake